Add ReporterBase
