+++ src/reporting3/format3-reporter.js
@@ -1,43 +1,13 @@
 const parser = require('./format3');
-const Builder = require('../reporting/builder');
-const ItemContainer = require('../reporting/item-container');
-const Output = require('../reporting/output');
+const ReporterBase = require('../reporting/reporter-base');
 
-class Format3Reporter {
+class Format3Reporter extends ReporterBase {
   constructor() {
-    /**
-     * @private
-     */
-    this._options = {
+    super(parser, {
       placeholder_mark: '#',
       bracket_open: '{',
       bracket_close: '}',
-    };
-  }
-
-  /**
-   * @param {ItemContainer} itemContainer
-   * @param {string} text
-   * @return {Output}
-   */
-  createReport(itemContainer, text) {
-    const parsed = parser.parse(text, this._options);
-    const reports = new Builder().build(parsed);
-
-    const output = new Output();
-
-    reports.forEach((report) => {
-      report.getCodes().forEach((code) => {
-        try {
-          const item = itemContainer.getItem(code);
-          const result = report.evaluate(item);
-
-          output.merge(result);
-        } catch (e) {}
-      });
     });
-
-    return output;
   }
 }
